Add integration test workflow so that we can run mount
