+++ test/shared_test.go
@@ -10,7 +10,6 @@ import (
 	"io/fs"
 	"net"
 	"os"
-	"path"
 	"path/filepath"
 	"sort"
 	"strings"
@@ -23,7 +22,6 @@ import (
 	"github.com/gadget-inc/dateilager/internal/pb"
 	util "github.com/gadget-inc/dateilager/internal/testutil"
 	"github.com/gadget-inc/dateilager/pkg/api"
-	"github.com/gadget-inc/dateilager/pkg/cached"
 	"github.com/gadget-inc/dateilager/pkg/client"
 	grpc_middleware "github.com/grpc-ecosystem/go-grpc-middleware"
 	"github.com/klauspost/compress/s2"
@@ -414,29 +412,6 @@ func createTestGRPCServer(tc util.TestCtx) (*bufconn.Listener, *grpc.Server, fun
 	}
 
 	return lis, s, getConn
-}
-
-func createTestCachedServer(tc util.TestCtx, tmpDir string) (*api.Cached, string, func()) {
-	cl, _, closeClient := createTestClient(tc)
-	_, grpcServer, _ := createTestGRPCServer(tc)
-
-	s := cached.CachedServer{
-		Grpc: grpcServer,
-	}
-
-	cached := tc.CachedApi(cl, path.Join(tmpDir, "cached", "staging"))
-	s.RegisterCached(cached)
-	s.RegisterCSI(cached)
-
-	socket := path.Join(tmpDir, "csi.sock")
-	endpoint := "unix://" + socket
-
-	go func() {
-		err := s.Serve(endpoint)
-		require.NoError(tc.T(), err, "CSI Server exited")
-	}()
-
-	return cached, endpoint, func() { closeClient(); s.Grpc.Stop() }
 }
 
 func createTestClient(tc util.TestCtx) (*client.Client, *api.Fs, func()) {
